refactor: My페이지 리팩토링
