+++ frontend/src/components/customer/my/index.ts
@@ -1,0 +1,3 @@
+export * from "./BottomList";
+export * from "./NutritionGoal";
+export * from "./MyInfo";
